+++ cql3/single_column_relation.hh
@@ -170,7 +170,7 @@ protected:
 
         auto term = to_term(to_receivers(*schema, column_def), *_value, db, schema->ks_name(), bound_names);
         auto r = ::make_shared<restrictions::single_column_restriction>(column_def);
-        r->expression = restrictions::make_column_op(&column_def, _relation_type, std::move(term));
+        r->expression = expr::make_column_op(&column_def, _relation_type, std::move(term));
         return r;
     }
 
@@ -180,7 +180,7 @@ protected:
         auto&& column_def = to_column_definition(*schema, *_entity);
         auto term = to_term(to_receivers(*schema, column_def), *_value, db, schema->ks_name(), bound_names);
         auto r = ::make_shared<restrictions::single_column_restriction>(column_def);
-        r->expression = restrictions::make_column_op(
+        r->expression = expr::make_column_op(
                 &column_def, is_key ? operator_type::CONTAINS_KEY : operator_type::CONTAINS, std::move(term));
         return r;
     }
